Experiment result with duktape
## Pros: Significantly smaller APK size
Just tried building a signed APK, and it’s 4.3mb. That’s like 10% the
size of the current approach.

## Cons: Regular expression handling too strict
Keep getting `D/Error: com.squareup.duktape.DuktapeException:
SyntaxError: invalid regexp escape` in many cases, such as when parsing
HTML. This is a deal breaker at the moment because it throws error for
many javascript libraries.

## Conclusion
Duktape approach seems to work fine for most cases, but when it breaks,
there’s no easy way to debug because the regular expression problem is
something we cannot control when we use multiple javascript libraries.
Going to try J2V8 next.