add action href for map pins (#149)
* improve error handling/logging in helper & docs

* Fix call method handling of data, also make it...

consistent across call, success and error handlers

* Add action for map pin selection

Action will be called when user selects a pin on the map. A new prop
default_action controls whether the GMaps default behaviour is called
(showing ping annotation) after the jasonette action is run.

* use per-pin actions

* Trigger action on map pin "info-window click"

This UX instead of action being on the initial pin click, so user sees
the infowindow for the pin and needs to lick on it to trigger the action.